Big fixes for batch processing. Deals with finding zero OpenSearch records and passes failed_list through functions.

+++ cci_tools/create_cci_stac_records/run_latest_cci_versions.py
@@ -46,31 +46,31 @@ def find_latest_versions(root_path):
                 latest_versions[dirpath] = os.path.join(dirpath, latest_ver)
             else:
                 print(f"WARNING: {latest_ver} and {latest_ver_num} don't match!")
-                latest_versions[dirpath] = None
+                raise ValueError(f"Data version directory mismatch: {latest_ver} and {latest_ver_num} don't match!")
 
     return latest_versions
 
 # Parse command line arguments using click
 @click.command()
 @click.argument('root_directory')
+@click.option('--owrite_list', is_flag=True, help='Overwrite existing latest version list.')
 
-def main(root_directory):
+def main(root_directory,owrite_list):
+    
+    run_latest(root_directory,owrite_list)
+
+def run_latest(root_directory,owrite_list):
+
     if root_directory.isnumeric():
         path_file='/home/users/dknappett/tools/cci-tools/cci_project_paths.txt'
         with open(path_file) as f:
             root_directory=[r.strip() for r in f.readlines()][int(root_directory)]
-
-    # Example usage
-    #root_directory = '/neodc/esacci/aerosol/data'
-    latest_dirs = find_latest_versions(root_directory)
-
-    #for parent, latest in latest_dirs.items():
-    #    print(f"{parent} → {latest}")
     
     # Extract project/ECV (i.e. after /neodc/esacci/)
     dir_split=root_directory.split('/')
     project=dir_split[3]
 
+    #output_dir=f'/gws/nopw/j04/esacci_portal/stac/stac_records/test/{project}'
     output_dir=f'/gws/nopw/j04/esacci_portal/stac/stac_records/lotus_jobs/{project}'
     if not os.path.exists(output_dir):
         try:
@@ -83,19 +83,31 @@ def main(root_directory):
 
     output_file=f"{output_dir}/latest_version_list.txt"
 
-    with open(output_file, "w") as file:
-        for parent, latest in latest_dirs.items():
-                file.write(latest + "\n")
-                print(f"{parent} → {latest}")
-    
-    print("")    
-    print(f"The latest version of CCI directories for the {project} project have been written to the following file:")
-    print(output_file)
-    print("")
+    if (not os.path.isfile(output_file) or owrite_list):
+        # Example usage
+        # root_directory = '/neodc/esacci/aerosol/data'
+        latest_dirs = find_latest_versions(root_directory)
+        values = [latest_dirs[key] for key in latest_dirs]
 
+        with open(output_file, "w") as file:
+            for parent, latest in latest_dirs.items():
+                    file.write(latest + "\n")
+                    print(f"{parent} → {latest}")
+        print("")    
+        print(f"The latest version of CCI directories for the {project} project have been written to the following file:")
+        print(output_file)
+        print("")
+
+    else:
+        print(f"Opening existing output file:")
+        print(output_file)
+        with open(output_file, 'r') as file:
+            values = [r.strip() for r in file.readlines()]
+            print(values)
+        
     # Loop over creation of STAC records
-    for parent, latest in latest_dirs.items():
-        create_cci_stac_records.create_stac(latest, output_dir, None)
+    for file_path in values:
+        create_cci_stac_records.create_stac(file_path, output_dir, None)
 
 if __name__ == "__main__":
     main()
